Proxies/ScraperAPI/requirements.txt
scrapy
scraperapi-sdk